feat: deletsdylan

+++ app/Services/UsuarioService.php
@@ -39,6 +39,14 @@ class UsuarioService
 
     public function delete(int $id)
     {
+    $usuario = $this->usuarioRepository->getById($id);
+
+        if (!$usuario) {
+            throw new ModelNotFoundException(
+            'Usuario no encontrado'
+        );
+        }
+
         return $this->usuarioRepository->delete($id);
     }
